CDEC-494: Rearrange Pos.DB.Block to prepare for consolidation
Consolidation of blocks into epoch files requires code out of Pos.DB.Block
and ends up having its functionality exported out of Pos.DB.Block. Therefore,
to avoid circular dependencies, Pos.DB.Block needs to be split up.
